feat: implement URL-based i18n routing with /en and /fr paths
- Add proxy.ts (Next.js 16): redirects / to /en or /fr based on
  Accept-Language, sets x-locale header for server-side html lang
- Add [locale]/layout.tsx: locale-specific metadata, hreflang alternates,
  canonical URLs, notFound() guard for invalid locales
- Move homepage and privacy page under [locale]/ for full SEO coverage
- Simplify LanguageProvider: locale derived from usePathname(), setLocale()
  navigates via router.push() — no more localStorage/navigator detection
- Update sitemap: /en, /fr, /en/privacy, /fr/privacy
- Update footer privacy link to /{locale}/privacy
