+++ src/app/[locale]/page.tsx
@@ -1,0 +1,29 @@
+import Navbar from '@/components/sections/Navbar';
+import Hero from '@/components/sections/Hero';
+import Services from '@/components/sections/Services';
+import Expertise from '@/components/sections/Expertise';
+import TechStackBar from '@/components/sections/TechStackBar';
+import Clients from '@/components/sections/Clients';
+import Testimonials from '@/components/sections/Testimonials';
+import Articles from '@/components/sections/Articles';
+import AppShowcase from '@/components/sections/AppShowcase';
+import ContactCTA from '@/components/sections/ContactCTA';
+import Footer from '@/components/sections/Footer';
+
+export default function Home() {
+	return (
+		<>
+			<Navbar />
+			<Hero />
+			<Services />
+			<Expertise />
+			<TechStackBar />
+			<Clients />
+			<Testimonials />
+			<Articles />
+			<AppShowcase />
+			<ContactCTA />
+			<Footer />
+		</>
+	);
+}
