+++ src/app/page.tsx
@@ -1,29 +1,5 @@
-import Navbar from '@/components/sections/Navbar';
-import Hero from '@/components/sections/Hero';
-import Services from '@/components/sections/Services';
-import Expertise from '@/components/sections/Expertise';
-import TechStackBar from '@/components/sections/TechStackBar';
-import Clients from '@/components/sections/Clients';
-import Testimonials from '@/components/sections/Testimonials';
-import Articles from '@/components/sections/Articles';
-import AppShowcase from '@/components/sections/AppShowcase';
-import ContactCTA from '@/components/sections/ContactCTA';
-import Footer from '@/components/sections/Footer';
+import { redirect } from 'next/navigation';
 
-export default function Home() {
-	return (
-		<>
-			<Navbar />
-			<Hero />
-			<Services />
-			<Expertise />
-			<TechStackBar />
-			<Clients />
-			<Testimonials />
-			<Articles />
-			<AppShowcase />
-			<ContactCTA />
-			<Footer />
-		</>
-	);
+export default function RootPage() {
+	redirect('/en');
 }
