+++ src/app/sitemap.ts
@@ -1,12 +1,33 @@
 import { MetadataRoute } from 'next'
 
+const BASE_URL = 'https://www.roget-concept.be'
+
 export default function sitemap(): MetadataRoute.Sitemap {
+  const lastModified = new Date()
   return [
     {
-      url: 'https://www.roget-concept.be/',
-      lastModified: new Date(),
+      url: `${BASE_URL}/en`,
+      lastModified,
       changeFrequency: 'monthly',
       priority: 1,
+    },
+    {
+      url: `${BASE_URL}/fr`,
+      lastModified,
+      changeFrequency: 'monthly',
+      priority: 1,
+    },
+    {
+      url: `${BASE_URL}/en/privacy`,
+      lastModified,
+      changeFrequency: 'yearly',
+      priority: 0.3,
+    },
+    {
+      url: `${BASE_URL}/fr/privacy`,
+      lastModified,
+      changeFrequency: 'yearly',
+      priority: 0.3,
     },
   ]
 }
