+++ src/i18n/LanguageProvider.tsx
@@ -3,10 +3,10 @@
 import {
 	createContext,
 	useContext,
-	useState,
 	useEffect,
 	type ReactNode,
 } from 'react';
+import { usePathname, useRouter } from 'next/navigation';
 import { en, type Translations } from '@/i18n/locales/en';
 import { fr } from '@/i18n/locales/fr';
 
@@ -30,30 +30,17 @@ export function useTranslation() {
 	return useContext(I18nContext);
 }
 
-function detectLocale(): Locale {
-	const saved = localStorage.getItem('locale') as Locale | null;
-	if (saved === 'en' || saved === 'fr') return saved;
-	if (navigator.language.toLowerCase().startsWith('fr')) return 'fr';
-	return 'en';
-}
-
 export function LanguageProvider({ children }: { children: ReactNode }) {
-	const [locale, setLocaleState] = useState<Locale>('en');
+	const pathname = usePathname();
+	const router = useRouter();
+	const locale: Locale = pathname?.startsWith('/fr') ? 'fr' : 'en';
 
 	useEffect(() => {
-		const detected = detectLocale();
-		document.documentElement.lang = detected;
-		if (detected !== 'en') {
-			// One-time browser locale detection on mount
-			// eslint-disable-next-line react-hooks/set-state-in-effect
-			setLocaleState(detected);
-		}
-	}, []);
+		document.documentElement.lang = locale;
+	}, [locale]);
 
 	const setLocale = (newLocale: Locale) => {
-		setLocaleState(newLocale);
-		localStorage.setItem('locale', newLocale);
-		document.documentElement.lang = newLocale;
+		router.push(`/${newLocale}`);
 	};
 
 	return (
